mods to previous q chart

+++ qna/questions/templatetags/questions_tags.py
@@ -12,6 +12,6 @@ def settings_value(name):
 @register.simple_tag
 def percentage(previous_question, answer):
       if previous_question.total_vote_count > 0:
-            return float(answer.votes.count()) / float(previous_question.total_vote_count) * 100
+            return int(float(answer.votes.count()) / float(previous_question.total_vote_count) * 100)
       else:
             return 0
